+++ OBBEdit/patch.py
@@ -2,28 +2,58 @@
 import datetime
 from hashlib import md5
 from io import BytesIO
-from os import makedirs, listdir, getcwd, sep
+from os import listdir, getcwd, sep
 from os.path import isdir, isfile, join as osjoin, dirname, relpath, splitext
 #from PIL import Image
 from struct import pack, unpack
 from zlib import compress, decompress
 
 # 3th party libraries
-from libraries.pyvz2nineteendo import LogError, blue_print, green_print, initialize, path_input, list_levels
+from libraries.pyvz2nineteendo import LogError, blue_print, green_print, initialize, mkdirs, path_input, list_levels
 from libraries.pyvz2rijndael import RijndaelCBC
 from libraries.pyvz2rton import JSONDecoder
 
 options = {
 # Default options
+	# ZIP options
+	"archiveExtensions": (
+		".rsb.smf",
+		
+		".1bsr",
+		".rsb1",
+		".rsb",
+		".obb",
+		
+		".pgsr",
+		".rsgp",
+		".rsg",
+		".rsg.smf"
+	),
+	"encodedExtensions": (
+		".rton",
+	),
+	"zipExtensions": (
+		".apk",
+		".zip"
+	),
+	"zipPacked": "",
+	"zipUnpacked": "",
+	"zipUnpackLevel": 1,
 	# SMF options
 	"smfExtensions": (
+		".apk",
+		".zip",
+
 		".rsb.smf",
 	),
 	"smfPacked": "",
 	"smfUnpacked": "",
-	"smfUnpackLevel": 1,
+	"smfUnpackLevel": 2,
 	# RSB options
 	"rsbExtensions": (
+		".apk",
+		".zip",
+
 		".rsb.smf",
 		
 		".1bsr",
@@ -34,7 +64,7 @@ options = {
 	"rsbPacked": "",
 	"rsbPatched": "",
 	"rsbUnpacked": "",
-	"rsbUnpackLevel": 2,
+	"rsbUnpackLevel": 3,
 	"rsgEndsWith": (),
 	"rsgEndsWithIgnore": True,
 	"rsgStartsWith": (
@@ -55,6 +85,9 @@ options = {
 	),
 	"pathStartsWithIgnore": False,
 	"rsgExtensions": (
+		".apk",
+		".zip",
+		
 		".rsb.smf",
 		
 		".1bsr",
@@ -70,21 +103,21 @@ options = {
 	"rsgPacked": "",
 	"rsgPatched": "",
 	"rsgUnpacked": "",
-	"rsgUnpackLevel": 7,
+	"rsgUnpackLevel": 8,
 	# Encryption options
 	"encryptedExtensions": (
 		".rton",
 	),
 	"encryptedPacked": "",
 	"encryptedUnpacked": "",
-	"encryptedUnpackLevel": 5,
+	"encryptedUnpackLevel": 6,
 	"encryptionKey": "00000000000000000000000000000000",
 	# RTON options
 	"comma": 0,
 	"doublePoint": 1,
 	"encodedPacked": "",
 	"encodedUnpacked": "",
-	"encodedUnpackLevel": 6,
+	"encodedUnpackLevel": 7,
 	"ensureAscii": False,
 	"indent": 4,
 	"repairFiles": False,
@@ -109,7 +142,7 @@ class SectionError(Exception):
 	pass
 def extend_to_4096(number):
 	return b"\0" * ((4096 - number) & 4095)
-def rsg_patch_data(RSG_NAME, file, pathout_data, patch, patchout, level):
+def rsg_patch_data(RSG_NAME, file, pathout_data, patch, patchout, unpack_level):
 # Patch RGSP file
 	HEADER = file.read(4)
 	VERSION = unpack("<I", file.read(4))[0]
@@ -133,7 +166,7 @@ def rsg_patch_data(RSG_NAME, file, pathout_data, patch, patchout, level):
 	INFO_LIMIT = INFO_OFFSET + INFO_SIZE
 
 	data = None
-	if level < 5:
+	if unpack_level < 6:
 		try:
 			patch_data = open(osjoin(patch, RSG_NAME + ".section"), "rb").read()
 			patch_length = len(patch_data)
@@ -150,7 +183,7 @@ def rsg_patch_data(RSG_NAME, file, pathout_data, patch, patchout, level):
 		
 	image_data = None
 	if DECOMPRESSED_IMAGE_DATA_SIZE != 0:
-		if level < 5:
+		if unpack_level < 6:
 			try:
 				patch_data = open(osjoin(patch, RSG_NAME + ".section2"), "rb").read()
 				patch_length = len(patch_data)
@@ -165,7 +198,7 @@ def rsg_patch_data(RSG_NAME, file, pathout_data, patch, patchout, level):
 		else: # Compressed files
 			image_data = bytearray(decompress(pathout_data[IMAGE_DATA_OFFSET: IMAGE_DATA_OFFSET + COMPRESSED_IMAGE_DATA_SIZE]))
 
-	if 4 < level:
+	if 5 < unpack_level:
 		DATA_DICT = {
 			"": {
 				"FILE_OFFSET": DECOMPRESSED_DATA_SIZE
@@ -225,7 +258,7 @@ def rsg_patch_data(RSG_NAME, file, pathout_data, patch, patchout, level):
 				FILE_INFO = DATA_DICT[DECODED_NAME]["FILE_INFO"]
 				if NAME_CHECK.startswith(pathStartsWith) and NAME_CHECK.endswith(pathEndsWith):
 					try:
-						if level < 7:
+						if unpack_level < 8:
 							file_name = osjoin(patch, DECODED_NAME)
 							patch_data = open(file_name, "rb").read()
 						elif NAME_CHECK[-5:] == ".rton":
@@ -234,7 +267,7 @@ def rsg_patch_data(RSG_NAME, file, pathout_data, patch, patchout, level):
 						else:
 							raise FileNotFoundError
 
-						if NAME_CHECK[-5:] == ".rton" and 5 < level and (overrideEncryption == 1 or overrideEncryption < 0 and data[FILE_OFFSET: FILE_OFFSET + 2] == b"\x10\0") and patch_data[0:2] != b"\x10\0":
+						if NAME_CHECK[-5:] == ".rton" and 6 < unpack_level and (overrideEncryption == 1 or overrideEncryption < 0 and data[FILE_OFFSET: FILE_OFFSET + 2] == b"\x10\0") and patch_data[0:2] != b"\x10\0":
 							patch_data = b'\x10\0' + rijndael_cbc.encrypt(patch_data)
 						
 						FILE_SIZE = len(patch_data)
@@ -261,7 +294,7 @@ def rsg_patch_data(RSG_NAME, file, pathout_data, patch, patchout, level):
 				FILE_INFO = IMAGE_DATA_DICT[DECODED_NAME]["FILE_INFO"]
 				if NAME_CHECK.startswith(pathStartsWith) and NAME_CHECK.endswith(pathEndsWith):
 					try:
-						#if level < 7:
+						#if unpack_level < 8:
 						file_name = osjoin(patch, DECODED_NAME)
 						patch_data = open(file_name, "rb").read()
 						#else:
@@ -301,7 +334,7 @@ def rsg_patch_data(RSG_NAME, file, pathout_data, patch, patchout, level):
 		pathout_data[DATA_OFFSET: IMAGE_DATA_OFFSET] = data
 		pathout_data[28:36] = pack("<I", COMPRESSED_DATA_SIZE) + pack("<I", DECOMPRESSED_DATA_SIZE)
 		pathout_data[40:44] = pack("<I", DATA_OFFSET + COMPRESSED_DATA_SIZE)
-		if level < 5:
+		if unpack_level < 6:
 			print("patched " + relpath(osjoin(patch, RSG_NAME + ".section"), patchout))
 		
 	if image_data != None:
@@ -319,12 +352,12 @@ def rsg_patch_data(RSG_NAME, file, pathout_data, patch, patchout, level):
 		
 		pathout_data[IMAGE_DATA_OFFSET:] = image_data
 		pathout_data[44:52] = pack("<I", COMPRESSED_IMAGE_DATA_SIZE) + pack("<I", DECOMPRESSED_IMAGE_DATA_SIZE)
-		if level < 5:
+		if unpack_level < 6:
 			print("patched " + relpath(osjoin(patch, RSG_NAME + ".section2"), patchout))
 	
 	pathout_data[16:20] = pack("<I", COMPRESSION_FLAGS)
 	return pathout_data
-def rsb_patch_data(file, pathout_data, patch, patchout, level):
+def rsb_patch_data(file, pathout_data, patch, patchout, unpack_level):
 	VERSION = unpack('<L', file.read(4))[0]
 
 	file.seek(4, 1)
@@ -414,14 +447,14 @@ def rsb_patch_data(file, pathout_data, patch, patchout, level):
 		RSG_CHECK = RSG_NAME.lower()
 		if RSG_CHECK.startswith(rsgStartsWith) and RSG_CHECK.endswith(rsgEndsWith):
 			try:
-				if level < 4:
+				if unpack_level < 5:
 					file_path = osjoin(patch, RSG_NAME + ".rsg")
 					subdata = bytearray(open(file_path, "rb").read())
 				else:
 					subdata = pathout_data[RSG_OFFSET: RSG_OFFSET + RSG_SIZE]
 					subdata[16:36] = pathout_data[info_start + 140:info_start + 160]
 					subdata[40:52] = pathout_data[info_start + 164:info_start + 176]
-					subdata = rsg_patch_data(RSG_NAME, BytesIO(subdata), subdata, patch, patchout, level)
+					subdata = rsg_patch_data(RSG_NAME, BytesIO(subdata), subdata, patch, patchout, unpack_level)
 				
 				subdata[:4] = b"pgsr"
 				subdata += extend_to_4096(len(subdata))
@@ -429,7 +462,7 @@ def rsb_patch_data(file, pathout_data, patch, patchout, level):
 				pathout_data[info_start + 132:info_start + 136] = pack("<I", len(subdata)) #RSG_SIZE
 				pathout_data[info_start + 140:info_start + 176] = subdata[16:36] + subdata[32:36] + subdata[40:52]
 				RSG_SHIFT += len(subdata) - RSG_SIZE
-				if level < 4:
+				if unpack_level < 5:
 					print("applied " + relpath(file_path, patchout))
 			except FileNotFoundError:
 				pass
@@ -437,94 +470,110 @@ def rsb_patch_data(file, pathout_data, patch, patchout, level):
 				error_message(e, " while patching " + RSG_NAME + ".rsg")
 		pathout_data[info_start + 128:info_start + 132] = pack("<I", RSG_OFFSET)
 	return pathout_data
-def file_to_folder(inp, out, patch, level, extensions, pathout, patchout):
+def archive_patch(file, out, patch, unpack_level, pathout, patchout):
+	HEADER = file.read(4)
+	COMPRESSED = HEADER == b"\xD4\xFE\xAD\xDE"
+	if COMPRESSED and 2 < unpack_level:
+		DECOMPRESSED_SIZE = unpack("<I", file.read(4))[0]
+		pathout_data = decompress(file.read())
+		name = file.name
+		file = BytesIO(pathout_data)
+		file.name = name
+		HEADER = file.read(4)
+
+	if HEADER == b"1bsr":
+		if not COMPRESSED:
+			pathout_data = HEADER + file.read()
+			file.seek(4)
+		
+		if unpack_level > 3:
+			pathout_data = rsb_patch_data(file, bytearray(pathout_data), patch, patchout, unpack_level)
+		if unpack_level < 4 or COMPRESSED:
+			tag, extension = splitext(out)
+			tag += ".tag" + extension
+			open(tag, "wb").write(md5(pathout_data).hexdigest().upper().encode() + b"\r\n")
+			green_print("wrote " + relpath(tag, pathout))
+			pathout_data = b"\xD4\xFE\xAD\xDE" + pack("<I", len(pathout_data)) + compress(pathout_data, level = 9)
+		
+		open(out, "wb").write(pathout_data)
+		green_print("wrote " + relpath(out, pathout))
+	elif HEADER == b"pgsr":
+		try:
+			pathout_data = bytearray(HEADER + file.read())
+			file.seek(0)
+			pathout_data = rsg_patch_data("data", file, pathout_data, patch, patchout, unpack_level)
+			open(out, "wb").write(pathout_data)
+			green_print("wrote " + relpath(out, pathout))
+		except Exception as e:
+			error_message(e, " while patching " + file.name)
+	elif not COMPRESSED:
+		warning_message("UNKNOWN HEADER (" + HEADER.hex() + ") in " + file.name)
+def file_encode(file, out, unpack_level, repack_level, pathout):
+	HEADER = file.read(1)
+	if HEADER == b"{":
+		if unpack_level > 7:
+			file.seek(0)
+			encoded_data = encode_root_object(file)
+			if repack_level > 6:
+				open(out, "wb").write(encoded_data)
+				print("wrote " + relpath(out, pathout))
+			else:
+				name = file.name
+				file = BytesIO(encoded_data)
+				file.name = name
+				HEADER = file.read(4)
+	else:
+		HEADER += file.read(3)
+
+	if HEADER == b"RTON":
+		if repack_level == 6:
+			open(out,"wb").write(b'\x10\0' + rijndael_cbc.encrypt(HEADER + file.read()))
+			print("wrote " + relpath(out, pathout))
+def file_to_folder(inp, out, patch, unpack_level, extensions, pathout, patchout):
 # Recursive file convert function
 	if isfile(inp):
 		try:
-			file = open(inp, "rb")
-			HEADER = file.read(4)
-			COMPRESSED = HEADER == b"\xD4\xFE\xAD\xDE" and 2 < level
-			if COMPRESSED:
-				DECOMPRESSED_SIZE = unpack("<I", file.read(4))[0]
-				pathout_data = decompress(file.read())
-				file = BytesIO(pathout_data)
-				file.name = inp
-				HEADER = file.read(4)
-
-			if HEADER == b"1bsr":
-				if not COMPRESSED:
-					pathout_data = HEADER + file.read()
-					file.seek(4)
-				
-				if level > 2:
-					pathout_data = rsb_patch_data(file, bytearray(pathout_data), patch, patchout, level)
-				if level < 3 or COMPRESSED:
-					tag, extension = splitext(out)
-					tag += ".tag" + extension
-					open(tag, "wb").write(md5(pathout_data).hexdigest().upper().encode() + b"\r\n")
-					green_print("wrote " + relpath(tag, pathout))
-					pathout_data = b"\xD4\xFE\xAD\xDE" + pack("<I", len(pathout_data)) + compress(pathout_data, level = 9)
-				
-				open(out, "wb").write(pathout_data)
-				green_print("wrote " + relpath(out, pathout))
-			elif HEADER == b"pgsr":
-				try:
-					pathout_data = bytearray(HEADER + file.read())
-					file.seek(0)
-					pathout_data = rsg_patch_data("data", file, pathout_data, patch, patchout, level)
-					open(out, "wb").write(pathout_data)
-					green_print("wrote " + relpath(out, pathout))
-				except Exception as e:
-					error_message(e, " while patching " + inp)
-			elif 2 < level:
-				warning_message("UNKNOWN 1BSR HEADER (" + HEADER.hex() + ") in " + inp)
+			with open(inp, "rb") as file:
+				archive_patch(file, out, patch, unpack_level, pathout, patchout)
 		except Exception as e:
-			error_message(e, " in " + inp + " pos " + repr(file.tell()), "Failed OBBPatch: ")
+			error_message(e, " in " + inp, "Failed OBBPatch: ")
 	elif isdir(inp):
-		makedirs(out, exist_ok = True)
-		makedirs(patch, exist_ok = True)
+		mkdirs(out)
+		mkdirs(patch)
 		for entry in sorted(listdir(inp)):
 			input_file = osjoin(inp, entry)
 			output_file = osjoin(out, entry)
 			patch_file = osjoin(patch, entry)
 			if isfile(input_file):
-				if level < 3:
+				if unpack_level < 4:
 					output_file += ".smf"
 				if entry.lower().endswith(extensions):
-					file_to_folder(input_file, output_file, splitext(patch_file)[0], level, extensions, pathout, patchout)
+					file_to_folder(input_file, output_file, splitext(patch_file)[0], unpack_level, extensions, pathout, patchout)
 			elif input_file != pathout and inp != patchout:
-				file_to_folder(input_file, output_file, patch_file, level, extensions, pathout, patchout)
-def conversion(inp, out, level, extensions, pathout):
+				file_to_folder(input_file, output_file, patch_file, unpack_level, extensions, pathout, patchout)
+def conversion(inp, out, unpack_level, repack_level, extensions, pathout):
 # Convert file
 	if isfile(inp):
 		try:
-			file = open(inp, "rb")
-			if file.read(4) == b"RTON":
-				if level < 7:
-					open(out,"wb").write(b'\x10\0' + rijndael_cbc.encrypt(b"RTON" + file.read()))
-					print("wrote " + relpath(out, pathout))
-			elif level > 6:
-				file.seek(0)
-				encoded_data = encode_root_object(file)
-				open(out, "wb").write(encoded_data)
-				print("wrote " + relpath(out, pathout))
+			with open(inp, "rb") as file:
+				file_encode(file, out, unpack_level, repack_level, pathout)
 		except Exception as e:
 			error_message(e, " in " + inp)
 	elif isdir(inp):
-		makedirs(out, exist_ok = True)
+		mkdirs(out)
 		for entry in listdir(inp):
 			input_file = osjoin(inp, entry)
 			output_file = osjoin(out, entry)
 			if isfile(input_file):
 				check = entry.lower()
-				if level > 6:
+				if unpack_level > 7:
 					output_file = output_file[:-5]
 					if "" == splitext(output_file)[1] and not check.startswith(RTONNoExtensions):
 						output_file += ".rton"
 				if check[-5:] == extensions:
-					conversion(input_file, output_file, level, extensions, pathout)
+					conversion(input_file, output_file, unpack_level, extensions, pathout)
 			elif input_file != pathout:
-				conversion(input_file, output_file, level, extensions, pathout)
+				conversion(input_file, output_file, unpack_level, extensions, pathout)
 # Start of the code
 try:
 	application_path = initialize()
@@ -535,19 +584,19 @@ try:
 	logerror.check_version(3, 9, 0)
 	
 	print("""\033[95m
-\033[1mOBBPatcher v1.2.0 (c) 2022 Nineteendo\033[22m
+\033[1mOBBPatcher v1.2.0b (c) 2022 Nineteendo\033[22m
 \033[1mCode based on:\033[22m Luigi Auriemma, Small Pea & 1Zulu
 \033[1mDocumentation:\033[22m Watto Studios, YingFengTingYu, TwinKleS-C & h3x4n1um
 \033[1mFollow PyVZ2 development:\033[22m \033[4mhttps://discord.gg/CVZdcGKVSw\033[24m
 \033[0m""")
 	options = logerror.load_template(options, osjoin(application_path, "options"), 2)
-	level_to_name = ["SPECIFY", "SMF", "RSB", "RSG", "SECTION", "ENCRYPTED", "ENCODED", "DECODED"]
+	level_to_name = ["SPECIFY", "ZIP", "SMF", "RSB", "RSG", "SECTION", "ENCRYPTED", "ENCODED", "DECODED"]
 	list_levels(level_to_name)
-	options["encodedUnpackLevel"] = input_level("ENCODED Unpack Level", 6, 7, options["encodedUnpackLevel"])
-	options["encryptedUnpackLevel"] = input_level("ENCRYPTED Unpack Level", 5, 6, options["encryptedUnpackLevel"])
-	options["rsgUnpackLevel"] = input_level("RSG/RSB/SMF Unpack Level", 3, 7, options["rsgUnpackLevel"])
-	options["rsbUnpackLevel"] = input_level("RSB/SMF Unpack Level", 2, 3, options["rsbUnpackLevel"])
-	options["smfUnpackLevel"] = input_level("SMF Unpack Level", 1, 2, options["smfUnpackLevel"])
+	options["encodedUnpackLevel"] = input_level("ENCODED Unpack Level", 7, 8, options["encodedUnpackLevel"])
+	options["encryptedUnpackLevel"] = input_level("ENCRYPTED Unpack Level", 6, 8, options["encryptedUnpackLevel"])
+	options["rsgUnpackLevel"] = input_level("RSG/RSB/SMF Unpack Level", 4, 8, options["rsgUnpackLevel"])
+	options["rsbUnpackLevel"] = input_level("RSB/SMF Unpack Level", 3, 4, options["rsbUnpackLevel"])
+	options["smfUnpackLevel"] = input_level("SMF Unpack Level", 2, 3, options["smfUnpackLevel"])
 	
 	if options["rsgStartsWithIgnore"]:
 		rsgStartsWith = ""
@@ -559,11 +608,11 @@ try:
 		rsgEndsWith = options["rsgEndsWith"]
 	
 	rijndael_cbc = RijndaelCBC(str.encode(options["encryptionKey"]), 24)
-	if 7 >= options["rsgUnpackLevel"] > 3:
+	if 8 >= options["rsgUnpackLevel"] > 4:
 		list_levels(["SPECIFY", "DEFAULT", "DISABLE", "ENABLE"])
 		overrideDataCompression = 2 * (input_level("Compress Data Override", 1, 3, options["overrideDataCompression"]) - 2)
 		overrideImageDataCompression = input_level("Compress Image Data Override", 1, 3, options["overrideImageDataCompression"]) - 2
-	if 7 >= options["rsgUnpackLevel"] > 5:
+	if 8 >= options["rsgUnpackLevel"] > 6:
 		overrideEncryption = input_level("Encrypt Override", 1, 3, options["overrideEncryption"]) - 2
 	
 	if options["pathEndsWithIgnore"]:
@@ -578,19 +627,19 @@ try:
 	encode_root_object = JSONDecoder().encode_root_object
 
 	blue_print("\nWorking directory: " + getcwd())
-	if 7 >= options["encodedUnpackLevel"] > 6:
+	if 8 >= options["encodedUnpackLevel"] > 7:
 		encoded_input = path_input("ENCODED " + level_to_name[options["encodedUnpackLevel"]] + " Input file or directory", options["encodedUnpacked"])
 		if isfile(encoded_input):
 			encoded_output = path_input("ENCODED Output file", options["encodedPacked"])
 		else:
 			encoded_output = path_input("ENCODED Output directory", options["encodedPacked"])
-	if 6 >= options["encryptedUnpackLevel"] > 5:
+	if 7 >= options["encryptedUnpackLevel"] > 6:
 		encrypted_input = path_input("ENCRYPTED " + level_to_name[options["encryptedUnpackLevel"]] + " Input file or directory", options["encryptedUnpacked"])
 		if isfile(encrypted_input):
 			encrypted_output = path_input("ENCRYPTED Output file", options["encryptedPacked"])
 		else:
 			encrypted_output = path_input("ENCRYPTED Output directory", options["encryptedPacked"])
-	if 7 >= options["rsgUnpackLevel"] > 3:
+	if 8 >= options["rsgUnpackLevel"] > 4:
 		rsg_input = path_input("RSG/RSB/SMF Input file or directory", options["rsgPacked"])
 		if isfile(rsg_input):
 			rsg_output = path_input("RSG/RSB/SMF Modded file", options["rsgPatched"])
@@ -598,7 +647,7 @@ try:
 			rsg_output = path_input("RSG/RSB/SMF Modded directory", options["rsgPatched"])
 		rsg_patch = path_input("RSG/RSB/SMF " + level_to_name[options["rsgUnpackLevel"]] + " Patch directory", options["rsgUnpacked"])
 	
-	if 3 >= options["rsbUnpackLevel"] > 2:
+	if 4 >= options["rsbUnpackLevel"] > 3:
 		rsb_input = path_input("RSB/SMF Input file or directory", options["rsbPacked"])
 		if isfile(rsb_input):
 			rsb_output = path_input("RSB/SMF Modded file", options["rsbPatched"])
@@ -606,7 +655,7 @@ try:
 			rsb_output = path_input("RSB/SMF Modded directory", options["rsbPatched"])
 		rsb_patch = path_input("RSB/SMF " + level_to_name[options["rsbUnpackLevel"]] + " Patch directory", options["rsbUnpacked"])
 	
-	if 2 >= options["smfUnpackLevel"] > 1:
+	if 3 >= options["smfUnpackLevel"] > 2:
 		smf_input = path_input("SMF " + level_to_name[options["smfUnpackLevel"]] + " Input file or directory", options["smfUnpacked"])
 		if isfile(smf_input):
 			smf_output = path_input("SMF Output file", options["smfPacked"])
@@ -615,15 +664,15 @@ try:
 
 	# Start file_to_folder
 	start_time = datetime.datetime.now()
-	if 7 >= options["encodedUnpackLevel"] > 6:
-		conversion(encoded_input, encoded_output, options["encodedUnpackLevel"], ".json", dirname(encoded_output))
-	if 6 >= options["encryptedUnpackLevel"] > 5:
-		conversion(encrypted_input, encrypted_output, options["encryptedUnpackLevel"], ".rton", dirname(encrypted_output))
-	if 7 >= options["rsgUnpackLevel"] > 3:
+	if 8 >= options["encodedUnpackLevel"] > 7:
+		conversion(encoded_input, encoded_output, options["encodedUnpackLevel"], 7, ".json", dirname(encoded_output))
+	if 7 >= options["encryptedUnpackLevel"] > 6:
+		conversion(encrypted_input, encrypted_output, options["encryptedUnpackLevel"], 6, ".rton", dirname(encrypted_output))
+	if 8 >= options["rsgUnpackLevel"] > 4:
 		file_to_folder(rsg_input, rsg_output, rsg_patch, options["rsgUnpackLevel"], options["rsgExtensions"], dirname(rsg_output), rsg_patch)
-	if 3 >= options["rsbUnpackLevel"] > 2:
+	if 4 >= options["rsbUnpackLevel"] > 3:
 		file_to_folder(rsb_input, rsb_output, rsb_patch, options["rsbUnpackLevel"], options["rsbExtensions"], dirname(rsb_output), rsb_patch)
-	if 2 >= options["smfUnpackLevel"] > 1:
+	if 3 >= options["smfUnpackLevel"] > 2:
 		file_to_folder(smf_input, smf_output, smf_output, options["smfUnpackLevel"], options["rsbExtensions"], dirname(smf_output), dirname(smf_output))
 
 	logerror.finish_program("finished patching in", start_time)