+++ OBBEdit/unpack.py
@@ -2,28 +2,59 @@
 import datetime
 
 from io import BytesIO
-from os import makedirs, listdir, getcwd, sep
+from os import listdir, getcwd, sep
 from os.path import isdir, isfile, join as osjoin, dirname, relpath, splitext
 #from PIL import Image
 from struct import unpack
+from zipfile import ZipFile
 from zlib import decompress
 
 # 3th party libraries
-from libraries.pyvz2nineteendo import LogError, blue_print, initialize, path_input, list_levels
+from libraries.pyvz2nineteendo import LogError, blue_print, green_print, initialize, mkdirs, path_input, list_levels
 from libraries.pyvz2rijndael import RijndaelCBC
 from libraries.pyvz2rton import RTONDecoder
 
 options = {
 # Default options
+	# ZIP options
+	"archiveExtensions": (
+		".rsb.smf",
+		
+		".1bsr",
+		".rsb1",
+		".rsb",
+		".obb",
+		
+		".pgsr",
+		".rsgp",
+		".rsg",
+		".rsg.smf"
+	),
+	"encodedExtensions": (
+		".rton",
+	),
+	"zipExtensions": (
+		".apk",
+		".zip"
+	),
+	"zipPacked": "",
+	"zipUnpacked": "",
+	"zipUnpackLevel": 1,
 	# SMF options
 	"smfExtensions": (
+		".apk",
+		".zip",
+
 		".rsb.smf",
 	),
 	"smfPacked": "",
 	"smfUnpacked": "",
-	"smfUnpackLevel": 1,
+	"smfUnpackLevel": 2,
 	# RSB options
 	"rsbExtensions": (
+		".apk",
+		".zip",
+
 		".rsb.smf",
 		
 		".1bsr",
@@ -34,7 +65,7 @@ options = {
 	"rsbPacked": "",
 	"rsbPatched": "",
 	"rsbUnpacked": "",
-	"rsbUnpackLevel": 2,
+	"rsbUnpackLevel": 3,
 	"rsgEndsWith": (),
 	"rsgEndsWithIgnore": True,
 	"rsgStartsWith": (
@@ -55,6 +86,9 @@ options = {
 	),
 	"pathStartsWithIgnore": False,
 	"rsgExtensions": (
+		".apk",
+		".zip",
+		
 		".rsb.smf",
 		
 		".1bsr",
@@ -70,21 +104,21 @@ options = {
 	"rsgPacked": "",
 	"rsgPatched": "",
 	"rsgUnpacked": "",
-	"rsgUnpackLevel": 7,
+	"rsgUnpackLevel": 8,
 	# Encryption options
 	"encryptedExtensions": (
 		".rton",
 	),
 	"encryptedPacked": "",
 	"encryptedUnpacked": "",
-	"encryptedUnpackLevel": 5,
+	"encryptedUnpackLevel": 6,
 	"encryptionKey": "00000000000000000000000000000000",
 	# RTON options
 	"comma": 0,
 	"doublePoint": 1,
 	"encodedPacked": "",
 	"encodedUnpacked": "",
-	"encodedUnpackLevel": 6,
+	"encodedUnpackLevel": 7,
 	"ensureAscii": False,
 	"indent": 4,
 	"repairFiles": False,
@@ -195,8 +229,8 @@ options = {
 # 	#149: XRGB8888_A8,
 # 	#150: ETC1_RGB_A_Palette
 # }
-#def rsg_extract(RSG_NAME, RSG_OFFSET, IMAGE_FORMATS, image_decoders, file, out, pathout, level):
-def rsg_extract(RSG_NAME, file, pathout_data, out, pathout, level):
+#def rsg_extract(RSG_NAME, RSG_OFFSET, IMAGE_FORMATS, image_decoders, file, out, pathout, unpack_level):
+def rsg_extract(RSG_NAME, file, pathout_data, out, pathout, unpack_level):
 	try:
 		HEADER = file.read(4)
 		VERSION = unpack("<I", file.read(4))[0]
@@ -231,7 +265,7 @@ def rsg_extract(RSG_NAME, file, pathout_data, out, pathout, level):
 			else: # Compressed files
 				image_data = bytearray(decompress(pathout_data[IMAGE_DATA_OFFSET: IMAGE_DATA_OFFSET + COMPRESSED_IMAGE_DATA_SIZE]))
 		
-		if level < 5:
+		if unpack_level < 6:
 			if COMPRESSION_FLAGS & 2 == 0 or COMPRESSED_DATA_SIZE != 0:
 				file_path = osjoin(out, RSG_NAME + ".section")
 				open(file_path, "wb").write(data)
@@ -276,15 +310,15 @@ def rsg_extract(RSG_NAME, file, pathout_data, out, pathout, level):
 					else:
 						file_data = data[FILE_OFFSET: FILE_OFFSET + FILE_SIZE]
 					
-					if NAME_CHECK[-5:] == ".rton" and file_data[:2] == b"\x10\0" and 5 < level:
+					if NAME_CHECK[-5:] == ".rton" and file_data[:2] == b"\x10\0" and 6 < unpack_level:
 						file_data = rijndael_cbc.decrypt(file_data[2:])
 
-					if NAME_CHECK[-5:] == ".rton" and 6 == level and file_data[:4] != b"RTON":
+					if NAME_CHECK[-5:] == ".rton" and 7 == unpack_level and file_data[:4] != b"RTON":
 						warning_message("No RTON " + file.name + ":" + DECODED_NAME)
 					else:
 						file_path = osjoin(out, DECODED_NAME)
-						makedirs(dirname(file_path), exist_ok = True)
-						if level > 6:
+						mkdirs(dirname(file_path))
+						if unpack_level > 7:
 							if NAME_CHECK[-5:] == ".rton":
 								try:
 									file_path = osjoin(out, DECODED_NAME[:-5] + ".JSON")
@@ -321,8 +355,8 @@ def rsg_extract(RSG_NAME, file, pathout_data, out, pathout, level):
 	except Exception as e:
 		error_message(e, " while extracting " + file.name)
 
-#def rsb_extract(file, out, level, image_decoders, pathout):
-def rsb_extract(file, pathout_data, out, level, pathout):
+#def rsb_extract(file, out, unpack_level, image_decoders, pathout):
+def rsb_extract(file, pathout_data, out, unpack_level, pathout):
 	VERSION = unpack('<L', file.read(4))[0]
 
 	file.seek(4, 1)
@@ -404,99 +438,154 @@ def rsb_extract(file, pathout_data, out, level, pathout):
 			subdata[:4] = b"pgsr"
 			subdata[16:36] = pathout_data[info_start + 140:info_start + 160]
 			subdata[40:52] = pathout_data[info_start + 164:info_start + 176]
-			if level < 4:
+			if unpack_level < 5:
 				open(osjoin(out, RSG_NAME + ".rsg"), "wb").write(subdata)
 				print("wrote " + relpath(osjoin(out, RSG_NAME + ".rsg"), pathout))
 			else:
 				subfile = BytesIO(subdata)
 				subfile.name = file.name + ":" + RSG_NAME
-				rsg_extract(RSG_NAME, subfile, subdata, out, pathout, level)
-				#rsg_extract(RSG_NAME, RSG_OFFSET, TEXTURE_FORMATS[IMAGE_ID:IMAGE_ID + IMAGE_ENTRIES], image_decoders, file, out, pathout, level)
-def file_to_folder(inp, out, level, extensions, pathout):
+				rsg_extract(RSG_NAME, subfile, subdata, out, pathout, unpack_level)
+				#rsg_extract(RSG_NAME, RSG_OFFSET, TEXTURE_FORMATS[IMAGE_ID:IMAGE_ID + IMAGE_ENTRIES], image_decoders, file, out, pathout, unpack_level)
+def archive_extract(file, out, unpack_level, pathout, allow_copy):
+	HEADER = file.read(4)
+	COMPRESSED = HEADER == b"\xD4\xFE\xAD\xDE"
+	if COMPRESSED:
+		if unpack_level > 2:
+			DECOMPRESSED_SIZE = unpack("<I", file.read(4))[0]
+			pathout_data = decompress(file.read())
+			if unpack_level > 3:
+				name = file.name
+				file = BytesIO(pathout_data)
+				file.name = name
+				HEADER = file.read(4)
+			else:
+				open(out, "wb").write(pathout_data)
+				print("wrote " + relpath(out, pathout))
+		elif allow_copy:
+			open(out, "wb").write(HEADER + file.read())
+			print("wrote " + relpath(out, pathout))
+	if HEADER == b"1bsr":
+		if unpack_level > 3:
+			if not COMPRESSED:
+				pathout_data = HEADER + file.read()
+				file.seek(4)
+			
+			# if file.[-4:] == ".obb":
+			# 	image_decoders = obb_image_decoders
+			# else:
+			# 	image_decoders = rsb_image_decoders
+			mkdirs(out)
+			rsb_extract(file, bytearray(pathout_data), out, unpack_level, pathout)
+			green_print("extracted " + relpath(out, pathout))
+			#rsb_extract(file, out, unpack_level, image_decoders, pathout)
+		elif allow_copy:
+			open(out, "wb").write(HEADER + file.read())
+			print("wrote " + relpath(out, pathout))
+	elif HEADER == b"pgsr":
+		if unpack_level > 4:
+			pathout_data = HEADER + file.read()
+			file.seek(0)
+			mkdirs(out)
+			rsg_extract("data", file, pathout_data, out, pathout, unpack_level)
+			green_print("extracted " + relpath(out, pathout))
+			#rsg_extract("data", 0, [], {} file, out, pathout, unpack_level)
+		elif allow_copy:
+			open(out, "wb").write(HEADER + file.read())
+			print("wrote " + relpath(out, pathout))
+	elif not COMPRESSED:
+		warning_message("UNKNOWN HEADER (" + HEADER.hex() + ") in " + file.name)
+def file_decode(file, out, unpack_level, pathout, allow_copy):
+	HEADER = file.read(2)
+	ENCRYPTED = HEADER == b"\x10\0"
+	if ENCRYPTED:
+		if unpack_level > 6:
+			file_data = rijndael_cbc.decrypt(file.read())
+			if file_data[:4] != b"RTON":
+				warning_message("No RTON " + file.name)
+			elif unpack_level > 7:
+				name = file.name
+				file = BytesIO(file_data)
+				file.name = name
+				HEADER = file.read(4)
+			else:
+				open(out,"wb").write(file_data)
+				print("wrote " + relpath(out, pathout))
+		elif allow_copy:
+			open(out, "wb").write(HEADER + file.read())
+			print("wrote " + relpath(out, pathout))
+	else:
+		HEADER += file.read(2)
+
+	if HEADER == b"RTON":
+		if unpack_level > 7:
+			data = parse_root_object(file)
+			open(out, "wb").write(data)
+			print("wrote " + relpath(out, pathout))
+		elif allow_copy:
+			open(out, "wb").write(HEADER + file.read())
+			print("wrote " + relpath(out, pathout))
+	elif file.name.lower()[-5:] != ".json" and not ENCRYPTED:
+		warning_message("UNKNOWN RTON HEADER (" + HEADER.hex() + ") in " + file.name)
+def file_to_folder(inp, out, unpack_level, extensions, pathout):
 # Recursive file convert function
 	if isfile(inp):
 		try:
-			file = open(inp, "rb")
-			HEADER = file.read(4)
-			COMPRESSED = HEADER == b"\xD4\xFE\xAD\xDE"
-			if COMPRESSED:
-				DECOMPRESSED_SIZE = unpack("<I", file.read(4))[0]
-				pathout_data = decompress(file.read())
-				if level < 3:
-					open(out, "wb").write(pathout_data)
-					print("wrote " + relpath(out, pathout))
-				else:
-					file = BytesIO(pathout_data)
-					file.name = inp
-					HEADER = file.read(4)
-			if HEADER == b"1bsr":
-				if not COMPRESSED:
-					pathout_data = HEADER + file.read()
-					file.seek(4)
-				
-				# if file.[-4:] == ".obb":
-				# 	image_decoders = obb_image_decoders
-				# else:
-				# 	image_decoders = rsb_image_decoders
-				makedirs(out, exist_ok = True)
-				rsb_extract(file, bytearray(pathout_data), out, level, pathout)
-				#rsb_extract(file, out, level, image_decoders, pathout)
-			elif HEADER == b"pgsr":
-				pathout_data = HEADER + file.read()
-				makedirs(out, exist_ok = True)
-				file.seek(0)
-				rsg_extract("data", file, pathout_data, out, pathout, level)
-				#rsg_extract("data", 0, [], {} file, out, pathout, level)
-			elif 2 < level:
-				warning_message("UNKNOWN 1BSR HEADER (" + HEADER.hex() + ") in " + inp)
+			if inp.lower().endswith((".apk", ".zip")):
+				with ZipFile(inp, 'r') as zipObj:
+					for name in zipObj.namelist():
+						if name.lower().endswith(options["archiveExtensions"]):
+							file_name = osjoin(out, name)
+							mkdirs(dirname(file_name))
+							with zipObj.open(name, "r") as file:
+								file.name = inp + ":" + name
+								archive_extract(file, file_name, unpack_level, pathout, True)
+						elif name.lower().endswith(options["encodedExtensions"]) and unpack_level > 5:
+							file_name = osjoin(out, name)
+							mkdirs(dirname(file_name))
+							with zipObj.open(name, "r") as file:
+								file.name = inp + ":" + name
+								file_decode(file, file_name, unpack_level, pathout, True)
+			else:
+				with open(inp, "rb") as file:
+					archive_extract(file, out, unpack_level, pathout, False)
 		except Exception as e:
-			error_message(e, " in " + inp + " pos " + repr(file.tell()), "Failed OBBUnpack: ")
+			error_message(e, " in " + inp, "Failed OBBUnpack: ")
 	elif isdir(inp):
-		makedirs(out, exist_ok = True)
+		mkdirs(out)
 		for entry in sorted(listdir(inp)):
 			input_file = osjoin(inp, entry)
 			output_file = osjoin(out, entry)
 			if isfile(input_file):
 				if entry.lower().endswith(extensions):
-					file_to_folder(input_file, splitext(output_file)[0], level, extensions, pathout)
+					file_to_folder(input_file, splitext(output_file)[0], unpack_level, extensions, pathout)
 			elif input_file != pathout:
-				file_to_folder(input_file, output_file, level, extensions, pathout)
-def conversion(inp, out, level, extensions, noextensions, pathout):
+				file_to_folder(input_file, output_file, unpack_level, extensions, pathout)
+def conversion(inp, out, unpack_level, extensions, noextensions, pathout):
 # Recursive file convert function
 	if isfile(inp):
 		try:
+			file = BytesIO()
 			file = open(inp, "rb")
-			HEADER = file.read(2)
-			if HEADER == b"\x10\0":
-				if level < 7:
-					open(out,"wb").write(rijndael_cbc.decrypt(file.read()))
-					print("wrote " + relpath(out, pathout))
-			else:
-				HEADER += file.read(2)
-				if HEADER == b"RTON":
-					if level > 6:
-						data = parse_root_object(file)
-						open(out, "wb").write(data)
-						print("wrote " + relpath(out, pathout))
-				elif inp.lower()[-5:] != ".json":
-					warning_message("UNKNOWN RTON HEADER (" + HEADER.hex() + ") in " + inp)
+			file_decode(file, out, unpack_level, pathout, False)
+			file.close()
 		except Exception as e:
 			error_message(e, " in " + inp + " pos " + repr(file.tell()))
+			file.close()
 	elif isdir(inp):
-		makedirs(out, exist_ok = True)
+		mkdirs(out)
 		for entry in listdir(inp):
 			input_file = osjoin(inp, entry)
 			output_file = osjoin(out, entry)
 			if isfile(input_file):
 				check = entry.lower()
-				if level > 6:
+				if unpack_level > 7:
 					if check[-5:] == ".rton":
 						output_file = output_file[:-5]
 					output_file += ".json"
 				if check.endswith(extensions) or check.startswith(noextensions):
-					conversion(input_file, output_file, level, extensions, noextensions, pathout)
+					conversion(input_file, output_file, unpack_level, extensions, noextensions, pathout)
 			elif input_file != pathout:
-				conversion(input_file, output_file, level, extensions, noextensions, pathout)
+				conversion(input_file, output_file, unpack_level, extensions, noextensions, pathout)
 # Start of the code
 try:
 	application_path = initialize()
@@ -507,19 +596,20 @@ try:
 	logerror.check_version(3, 9, 0)
 
 	print("""\033[95m
-\033[1mOBBUnpacker v1.2.0 (c) 2022 Nineteendo\033[22m
+\033[1mOBBUnpacker v1.2.0b (c) 2022 Nineteendo\033[22m
 \033[1mCode based on:\033[22m Luigi Auriemma, Small Pea & 1Zulu
 \033[1mDocumentation:\033[22m Watto Studios, YingFengTingYu, TwinKleS-C & h3x4n1um
 \033[1mFollow PyVZ2 development:\033[22m \033[4mhttps://discord.gg/CVZdcGKVSw\033[24m
 \033[0m""")
 	options = logerror.load_template(options, osjoin(application_path, "options"), 1)
-	level_to_name = ["SPECIFY", "SMF", "RSB", "RSG", "SECTION", "ENCRYPTED", "ENCODED", "DECODED"]
+	level_to_name = ["SPECIFY", "ZIP", "SMF", "RSB", "RSG", "SECTION", "ENCRYPTED", "ENCODED", "DECODED"]
 	list_levels(level_to_name)
-	options["smfUnpackLevel"] = input_level("SMF Unpack Level", 1, 2, options["smfUnpackLevel"])
-	options["rsbUnpackLevel"] = input_level("RSB/SMF Unpack Level", 2, 3, options["rsbUnpackLevel"])
-	options["rsgUnpackLevel"] = input_level("RSG/RSB/SMF Unpack Level", 3, 7, options["rsgUnpackLevel"])
-	options["encryptedUnpackLevel"] = input_level("ENCRYPTED Unpack Level", 5, 6, options["encryptedUnpackLevel"])
-	options["encodedUnpackLevel"] = input_level("ENCODED Unpack Level", 6, 7, options["encodedUnpackLevel"])
+	options["zipUnpackLevel"] = input_level("ZIP Unpack Level", 1, 2, options["zipUnpackLevel"])
+	options["smfUnpackLevel"] = input_level("SMF/ZIP Unpack Level", 2, 3, options["smfUnpackLevel"])
+	options["rsbUnpackLevel"] = input_level("RSB/SMF/ZIP Unpack Level", 3, 4, options["rsbUnpackLevel"])
+	options["rsgUnpackLevel"] = input_level("RSG/RSB/SMF/ZIP Unpack Level", 4, 8, options["rsgUnpackLevel"])
+	options["encryptedUnpackLevel"] = input_level("ENCRYPTED Unpack Level", 6, 8, options["encryptedUnpackLevel"])
+	options["encodedUnpackLevel"] = input_level("ENCODED/ENCRYPTED Unpack Level", 7, 8, options["encodedUnpackLevel"])
 
 	if options["rsgStartsWithIgnore"]:
 		rsgStartsWith = ""
@@ -563,42 +653,50 @@ try:
 	parse_root_object = RTONDecoder(comma, current_indent, doublePoint, ensureAscii, indent, repairFiles, sortKeys, sortValues, warning_message).parse_root_object
 	
 	blue_print("\nWorking directory: " + getcwd())
-	if 2 >= options["smfUnpackLevel"] > 1:
-		smf_input = path_input("SMF Input file or directory", options["smfPacked"])
-		if isfile(smf_input):
-			smf_output = path_input("SMF " + level_to_name[options["smfUnpackLevel"]] + " Output file", options["smfUnpacked"])
+	if 2 >= options["zipUnpackLevel"] > 1:
+		zip_input = path_input("ZIP Input file or directory", options["zipPacked"])
+		if isfile(zip_input):
+			zip_output = path_input("ZIP " + level_to_name[options["zipUnpackLevel"]] + " Output file", options["zipUnpacked"])
 		else:
-			smf_output = path_input("SMF " + level_to_name[options["smfUnpackLevel"]] + " Output directory", options["smfUnpacked"])
-	if 3 >= options["rsbUnpackLevel"] > 2:
-		rsb_input = path_input("RSB/SMF Input file or directory", options["rsbPacked"])
-		rsb_output = path_input("RSB/SMF " + level_to_name[options["rsbUnpackLevel"]] + " Output directory", options["rsbUnpacked"])
-	if 7 >= options["rsgUnpackLevel"] > 3:
-		rsg_input = path_input("RSG/RSB/SMF Input file or directory", options["rsgPacked"])
-		rsg_output = path_input("RSG/RSB/SMF " + level_to_name[options["rsgUnpackLevel"]] + " Output directory", options["rsgUnpacked"])
-	if 6 >= options["encryptedUnpackLevel"] > 5:
+			zip_output = path_input("ZIP " + level_to_name[options["zipUnpackLevel"]] + " Output directory", options["zipUnpacked"])
+	if 3 >= options["smfUnpackLevel"] > 2:
+		smf_input = path_input("SMF/ZIP Input file or directory", options["smfPacked"])
+		if isfile(smf_input):
+			smf_output = path_input("SMF/ZIP " + level_to_name[options["smfUnpackLevel"]] + " Output file", options["smfUnpacked"])
+		else:
+			smf_output = path_input("SMF/ZIP " + level_to_name[options["smfUnpackLevel"]] + " Output directory", options["smfUnpacked"])
+	if 4 >= options["rsbUnpackLevel"] > 3:
+		rsb_input = path_input("RSB/SMF/ZIP Input file or directory", options["rsbPacked"])
+		rsb_output = path_input("RSB/SMF/ZIP " + level_to_name[options["rsbUnpackLevel"]] + " Output directory", options["rsbUnpacked"])
+	if 8 >= options["rsgUnpackLevel"] > 4:
+		rsg_input = path_input("RSG/RSB/SMF/ZIP Input file or directory", options["rsgPacked"])
+		rsg_output = path_input("RSG/RSB/SMF/ZIP " + level_to_name[options["rsgUnpackLevel"]] + " Output directory", options["rsgUnpacked"])
+	if 8 >= options["encryptedUnpackLevel"] > 6:
 		encrypted_input = path_input("ENCRYPTED Input file or directory", options["encryptedPacked"])
 		if isfile(encrypted_input):
 			encrypted_output = path_input("ENCRYPTED " + level_to_name[options["encryptedUnpackLevel"]] + " Output file", options["encryptedUnpacked"])
 		else:
 			encrypted_output = path_input("ENCRYPTED " + level_to_name[options["encryptedUnpackLevel"]] + " Output directory", options["encryptedUnpacked"])
-	if 7 >= options["encodedUnpackLevel"] > 6:
-		encoded_input = path_input("ENCODED Input file or directory", options["encodedPacked"])
+	if 8 >= options["encodedUnpackLevel"] > 7:
+		encoded_input = path_input("ENCODED/ENCRYPTED Input file or directory", options["encodedPacked"])
 		if isfile(encoded_input):
-			encoded_output = path_input("ENCODED " + level_to_name[options["encodedUnpackLevel"]] + " Output file", options["encodedUnpacked"])
+			encoded_output = path_input("ENCODED/ENCRYPTED " + level_to_name[options["encodedUnpackLevel"]] + " Output file", options["encodedUnpacked"])
 		else:
-			encoded_output = path_input("ENCODED " + level_to_name[options["encodedUnpackLevel"]] + " Output directory", options["encodedUnpacked"])
+			encoded_output = path_input("ENCODED/ENCRYPTED " + level_to_name[options["encodedUnpackLevel"]] + " Output directory", options["encodedUnpacked"])
 
 	# Start file_to_folder
 	start_time = datetime.datetime.now()
-	if 2 >= options["smfUnpackLevel"] > 1:
+	if 2 >= options["zipUnpackLevel"] > 1:
+		file_to_folder(zip_input, zip_output, options["zipUnpackLevel"], options["zipExtensions"], dirname(zip_output))
+	if 3 >= options["smfUnpackLevel"] > 2:
 		file_to_folder(smf_input, smf_output, options["smfUnpackLevel"], options["smfExtensions"], dirname(smf_output))
-	if 3 >= options["rsbUnpackLevel"] > 2:
+	if 4 >= options["rsbUnpackLevel"] > 3:
 		file_to_folder(rsb_input, rsb_output, options["rsbUnpackLevel"], options["rsbExtensions"], rsb_output)
-	if 7 >= options["rsgUnpackLevel"] > 3:
+	if 8 >= options["rsgUnpackLevel"] > 4:
 		file_to_folder(rsg_input, rsg_output, options["rsgUnpackLevel"], options["rsgExtensions"], rsg_output)
-	if 6 >= options["encryptedUnpackLevel"] > 5:
+	if 8 >= options["encryptedUnpackLevel"] > 6:
 		conversion(encrypted_input, encrypted_output, options["encryptedUnpackLevel"], options["encryptedExtensions"], (), dirname(encrypted_output))
-	if 7 >= options["encodedUnpackLevel"] > 6:
+	if 8 >= options["encodedUnpackLevel"] > 7:
 		conversion(encoded_input, encoded_output, options["encodedUnpackLevel"], options["RTONExtensions"], options["RTONNoExtensions"], dirname(encoded_output))
 
 	logerror.finish_program("finished unpacking in", start_time)